SkeletalMesh pickups
Decided on a good way to do pickups utilizing skeletalmeshes. The "PickupMesh" can be set invisible and used only for collision purposes. The skeletal mesh with animations then is assigned as a child.

TODO: Make ground effect on the thruster realisted. Make TL default off (even though I love it and it's my baby)