New meshes and animation logic
Got it all working out the way I want it to, pistol is controlled by parameters in the animation blueprint vice animation montages. Can make timelines later if I so desire.

Still need to make a pistol projectile, it's still using the turret shell.

Need to make the pistol ignore the magazine load sphere when checking for overlap to attach the magazine.